requirements.txt
selenium
webdriver_manager
discord
pynacl
alive_progress
pytube
py-cord
openai-whisper